% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDecisionFeatures.R
\name{extractDecisionFeatures}
\alias{extractDecisionFeatures}
\title{Extracts both model-free and model-based features from the decision data}
\usage{
extractDecisionFeatures(decision_data)
}
\arguments{
\item{decision_data}{A clean decision data frame}
}
\value{
A data frame which merges both model-free and model-based parameter estimates
}
\description{
Extracts both model-free and model-based features from the decision data
}
