% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromEprimeSPSSExport.R
\name{getLevelsFromLotteryName}
\alias{getLevelsFromLotteryName}
\title{From a filename of risk-and-ambiguity lottery display, extract kind and level of lottery}
\usage{
getLevelsFromLotteryName(name)
}
\arguments{
\item{name}{Filename in the format kind_winColor_numericalLevel (e.g. ambig_red_25.png)}
}
\value{
A data frame with winProb, ambiguity, and winColor columns
}
\description{
From a filename of risk-and-ambiguity lottery display, extract kind and level of lottery
}
