% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromEprime.R
\name{importFromEprime}
\alias{importFromEprime}
\title{Import a single subject's R&A choice data from a text file exported by E-Prime.}
\usage{
importFromEprime(filename)
}
\arguments{
\item{filename}{A full file path to a .txt file exported by E-Prime}
}
\value{
A clean R&A data frame
}
\description{
Import a single subject's R&A choice data from a text file exported by E-Prime.
}
