% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFromEprimeSPSSExport.R
\name{importFromEprimeSPSSExport}
\alias{importFromEprimeSPSSExport}
\title{Import a single subject's R&A choice data from a text file exported by E-Prime.}
\usage{
importFromEprimeSPSSExport(filename, choiceColumnName = "choice",
  discardOriginalColumns = FALSE)
}
\arguments{
\item{filename}{A full file path to a .txt file exported by E-Prime as "SPSS and Statview file"}

\item{choiceColumnName}{The name of the E-Prime column in which the subject's choice is recorded}

\item{discardOriginalColumns}{Purge the additional E-Prime columns from the returned data frame?}
}
\value{
A clean R&A data frame
}
\description{
Import a single subject's R&A choice data from a text file exported by E-Prime.
}
