% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levylab.RNA.R
\docType{package}
\name{levylab.RNA}
\alias{levylab.RNA}
\alias{levylab.RNA-package}
\title{levylab.RNA: A package for analysing risk & ambiguity task data}
\description{
levylab.RNA: A package for analysing risk & ambiguity task data
}
\section{levylab.RNA functions}{

\itemize{
\item Import.
\item Setup.
\item Analysis.
}
}

