% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSetup.R
\name{setupAnalysisTemplates}
\alias{setupAnalysisTemplates}
\title{Copy the RMarkdown analysis templates into a directory}
\usage{
setupAnalysisTemplates(targetDirectory = "analysis/", includeOnly = NULL)
}
\arguments{
\item{targetDirectory}{A subdirectory in the current working
directory to which the analysis files will be copied}

\item{includeOnly}{A vector of RMarkdown filenames}
}
\value{
Outcome of file.copy
}
\description{
Copy the analysis templates that were installed with the package to
a targetDirectory which is, by default, the analysis/ folder in the
current working directory. (It will be created if it does not exist
already.) If includeOnly is provided, then only those scripts will
be copied.
}
