% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSetup.R
\name{setupDataDirectory}
\alias{setupDataDirectory}
\title{Create the data directory and subdirectories}
\usage{
setupDataDirectory(directory = NULL)
}
\arguments{
\item{directory}{The directory in which the data directories should be created}
}
\value{
Outcome of dir.create
}
\description{
Creates 'data/', 'data/raw', 'data/clean' and 'data/results' in the
directory provided or, by default, in the current working directory.
}
