% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDecisionFeatures.R
\name{getModelFreeEstimates}
\alias{getModelFreeEstimates}
\title{Estimates risk and ambiguity aversion based on a model-free comparison-based approach}
\usage{
getModelFreeEstimates(decision_data)
}
\arguments{
\item{decision_data}{A clean decision data frame}
}
\value{
A data frame with risk_proportion, ambiguity_proportion and relative_ambiguity_proportion
    parameters for each subject in decision_data.
}
\description{
Estimates risk and ambiguity aversion based on a model-free comparison-based approach
}
